Porting tile cache

+++ detourtilecache/src/main/java/org/recast4j/detour/tilecache/TileCacheLayer.java
@@ -1,10 +1,12 @@
 package org.recast4j.detour.tilecache;
 
 public class TileCacheLayer {
+	
 	TileCacheLayerHeader header;
 	int regCount;					///< Region count.
-	int[] heights;
-	int[] areas;
-	int[] cons;
-	int[] regs;
+	byte[] heights; // char 
+	byte[] areas;   // char
+	byte[] cons;    // char
+	byte[] regs;    // char
+	
 }
